Return proper ID for BiomeType
Rename MixinBiomeGen* -> MixinBiome*
