+++ src/main/java/org/spongepowered/common/mixin/core/world/biome/MixinBiomeMesa.java
@@ -26,32 +26,55 @@ package org.spongepowered.common.mixin.core.world.biome;
 
 import net.minecraft.world.World;
 import net.minecraft.world.biome.BiomeDecorator;
-import net.minecraft.world.biome.BiomeSavanna;
-import org.spongepowered.api.data.type.DoublePlantTypes;
+import net.minecraft.world.biome.BiomeMesa;
+import net.minecraft.world.chunk.ChunkPrimer;
 import org.spongepowered.api.util.weighted.VariableAmount;
-import org.spongepowered.api.world.gen.populator.DoublePlant;
+import org.spongepowered.api.world.gen.populator.Cactus;
 import org.spongepowered.api.world.gen.populator.Forest;
 import org.spongepowered.api.world.gen.type.BiomeTreeTypes;
+import org.spongepowered.asm.mixin.Final;
 import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.Shadow;
+import org.spongepowered.asm.mixin.injection.At;
+import org.spongepowered.asm.mixin.injection.Inject;
+import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 import org.spongepowered.common.world.biome.SpongeBiomeGenerationSettings;
+import org.spongepowered.common.world.gen.populators.MesaBiomeGenerationPopulator;
 
-@Mixin(BiomeSavanna.class)
-public abstract class MixinBiomeGenSavanna extends MixinBiomeGenBase {
+import java.util.Random;
+
+@Mixin(BiomeMesa.class)
+public abstract class MixinBiomeMesa extends MixinBiome {
+
+    @Shadow @Final private boolean brycePillars;
+    @Shadow @Final private boolean hasForest;
 
     @Override
     public void buildPopulators(World world, SpongeBiomeGenerationSettings gensettings) {
-        DoublePlant grass = DoublePlant.builder()
-                .type(DoublePlantTypes.GRASS, 1)
-                .perChunk(7 * 5)
-                .build();
-        gensettings.getPopulators().add(grass);
+        gensettings.getGenerationPopulators().add(new MesaBiomeGenerationPopulator(this.brycePillars, this.hasForest));
         super.buildPopulators(world, gensettings);
         BiomeDecorator theBiomeDecorator = this.theBiomeDecorator;
+        gensettings.getGroundCoverLayers().clear();
         gensettings.getPopulators().removeAll(gensettings.getPopulators(Forest.class));
         Forest.Builder forest = Forest.builder();
         forest.perChunk(VariableAmount.baseWithOptionalAddition(theBiomeDecorator.treesPerChunk, 1, 0.1));
         forest.type(BiomeTreeTypes.OAK.getPopulatorObject(), 1);
-        forest.type(BiomeTreeTypes.SAVANNA.getPopulatorObject(), 4);
         gensettings.getPopulators().add(0, forest.build());
+        gensettings.getPopulators().removeAll(gensettings.getPopulators(Cactus.class));
+        Cactus cactus = Cactus.builder()
+                .cactiPerChunk(VariableAmount.baseWithOptionalAddition(0,
+                        VariableAmount.baseWithRandomAddition(1, VariableAmount.baseWithOptionalAddition(2, 3, 0.25)), 0.4))
+                .build();
+        gensettings.getPopulators().add(cactus);
     }
+
+    /**
+     * Cancel the call to place the terrain blocks as this is instead handled
+     * through our custom genpop.
+     */
+    @Inject(method = "genTerrainBlocks(Lnet/minecraft/world/World;Ljava/util/Random;Lnet/minecraft/world/chunk/ChunkPrimer;IID)V", at = @At("HEAD") , cancellable = true)
+    public void genTerrainBlocks(World world, Random rand, ChunkPrimer chunk, int x, int z, double stoneNoise, CallbackInfo ci) {
+        ci.cancel();
+    }
+
 }
