+++ src/main/java/org/spongepowered/common/mixin/core/world/biome/MixinBiomeSavanna.java
@@ -25,15 +25,33 @@
 package org.spongepowered.common.mixin.core.world.biome;
 
 import net.minecraft.world.World;
-import net.minecraft.world.biome.BiomeVoid;
+import net.minecraft.world.biome.BiomeDecorator;
+import net.minecraft.world.biome.BiomeSavanna;
+import org.spongepowered.api.data.type.DoublePlantTypes;
+import org.spongepowered.api.util.weighted.VariableAmount;
+import org.spongepowered.api.world.gen.populator.DoublePlant;
+import org.spongepowered.api.world.gen.populator.Forest;
+import org.spongepowered.api.world.gen.type.BiomeTreeTypes;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.common.world.biome.SpongeBiomeGenerationSettings;
 
-@Mixin(BiomeVoid.class)
-public class MixinBiomeGenVoid extends MixinBiomeGenBase {
+@Mixin(BiomeSavanna.class)
+public abstract class MixinBiomeSavanna extends MixinBiome {
 
     @Override
     public void buildPopulators(World world, SpongeBiomeGenerationSettings gensettings) {
-        // no super call
+        DoublePlant grass = DoublePlant.builder()
+                .type(DoublePlantTypes.GRASS, 1)
+                .perChunk(7 * 5)
+                .build();
+        gensettings.getPopulators().add(grass);
+        super.buildPopulators(world, gensettings);
+        BiomeDecorator theBiomeDecorator = this.theBiomeDecorator;
+        gensettings.getPopulators().removeAll(gensettings.getPopulators(Forest.class));
+        Forest.Builder forest = Forest.builder();
+        forest.perChunk(VariableAmount.baseWithOptionalAddition(theBiomeDecorator.treesPerChunk, 1, 0.1));
+        forest.type(BiomeTreeTypes.OAK.getPopulatorObject(), 1);
+        forest.type(BiomeTreeTypes.SAVANNA.getPopulatorObject(), 4);
+        gensettings.getPopulators().add(0, forest.build());
     }
 }
