+++ src/main/resources/mixins.common.core.json
@@ -413,19 +413,19 @@
         "world.MixinWorldServerMulti",
         "world.MixinWorldSettings",
         "world.MixinWorldType",
-        "world.biome.MixinBiomeGenBase",
-        "world.biome.MixinBiomeGenDesert",
-        "world.biome.MixinBiomeGenEnd",
-        "world.biome.MixinBiomeGenForest",
-        "world.biome.MixinBiomeGenHell",
-        "world.biome.MixinBiomeGenHills",
-        "world.biome.MixinBiomeGenJungle",
-        "world.biome.MixinBiomeGenMesa",
-        "world.biome.MixinBiomeGenPlains",
-        "world.biome.MixinBiomeGenSavanna",
-        "world.biome.MixinBiomeGenSnow",
-        "world.biome.MixinBiomeGenTaiga",
-        "world.biome.MixinBiomeGenSwamp",
+        "world.biome.MixinBiome",
+        "world.biome.MixinBiomeDesert",
+        "world.biome.MixinBiomeEnd",
+        "world.biome.MixinBiomeForest",
+        "world.biome.MixinBiomeHell",
+        "world.biome.MixinBiomeHills",
+        "world.biome.MixinBiomeJungle",
+        "world.biome.MixinBiomeMesa",
+        "world.biome.MixinBiomePlains",
+        "world.biome.MixinBiomeSavanna",
+        "world.biome.MixinBiomeSnow",
+        "world.biome.MixinBiomeTaiga",
+        "world.biome.MixinBiomeSwamp",
         "world.biome.MixinBiomeProvider",
         "world.chunk.storage.MixinAnvilChunkLoader",
         "world.extent.MixinExtent",
